Extract workspace configuration APIs
