+++ Tests/QuillCodeParityTests/ParityWorkspaceModelGateTests.swift
@@ -237,6 +237,7 @@ final class ParityWorkspaceModelGateTests: QuillCodeParityTestCase {
 
     func testWorkspaceModelDelegatesConfigurationTransitions() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let configurationExtensionText = try Self.appSourceText(named: "WorkspaceModelConfiguration.swift")
         let configurationText = try Self.appSourceText(named: "WorkspaceConfigurationEngine.swift")
 
         XCTAssertTrue(configurationText.contains("struct WorkspaceConfigurationEngine"), "Workspace configuration transitions should live in a focused engine.")
@@ -246,11 +247,25 @@ final class ParityWorkspaceModelGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(configurationText.contains("static func normalizedCatalog"), "Catalog replacement should be directly testable.")
         XCTAssertTrue(configurationText.contains("static func applySettings"), "Settings application should be directly testable.")
         XCTAssertTrue(configurationText.contains("static func syncThread"), "Selected-thread config syncing should be directly testable.")
-        XCTAssertTrue(modelText.contains("WorkspaceConfigurationEngine.setModel"), "WorkspaceModel should delegate model selection.")
-        XCTAssertTrue(modelText.contains("WorkspaceConfigurationEngine.setMode"), "WorkspaceModel should delegate mode selection.")
-        XCTAssertTrue(modelText.contains("WorkspaceConfigurationEngine.toggleFavorite"), "WorkspaceModel should delegate favorite mutation.")
-        XCTAssertTrue(modelText.contains("WorkspaceConfigurationEngine.normalizedCatalog"), "WorkspaceModel should delegate catalog normalization.")
-        XCTAssertTrue(modelText.contains("WorkspaceConfigurationEngine.applySettings"), "WorkspaceModel should delegate settings application.")
+        XCTAssertTrue(configurationExtensionText.contains("public func setMode"), "Mode selection should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func setModel"), "Model selection should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func toggleModelFavorite"), "Favorite model mutation should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func setModelCatalog"), "Catalog replacement should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func applySettings"), "Settings application should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func applyRuntime"), "Runtime application should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("public func setAgentStatus"), "Agent status overrides should live in the focused configuration extension.")
+        XCTAssertTrue(configurationExtensionText.contains("WorkspaceConfigurationEngine.setModel"), "WorkspaceModelConfiguration should delegate model selection.")
+        XCTAssertTrue(configurationExtensionText.contains("WorkspaceConfigurationEngine.setMode"), "WorkspaceModelConfiguration should delegate mode selection.")
+        XCTAssertTrue(configurationExtensionText.contains("WorkspaceConfigurationEngine.toggleFavorite"), "WorkspaceModelConfiguration should delegate favorite mutation.")
+        XCTAssertTrue(configurationExtensionText.contains("WorkspaceConfigurationEngine.normalizedCatalog"), "WorkspaceModelConfiguration should delegate catalog normalization.")
+        XCTAssertTrue(configurationExtensionText.contains("WorkspaceConfigurationEngine.applySettings"), "WorkspaceModelConfiguration should delegate settings application.")
+        XCTAssertFalse(modelText.contains("public func setMode"), "WorkspaceModel.swift should not own mode selection API bodies.")
+        XCTAssertFalse(modelText.contains("public func setModel"), "WorkspaceModel.swift should not own model selection API bodies.")
+        XCTAssertFalse(modelText.contains("public func toggleModelFavorite"), "WorkspaceModel.swift should not own favorite model API bodies.")
+        XCTAssertFalse(modelText.contains("public func setModelCatalog"), "WorkspaceModel.swift should not own model catalog API bodies.")
+        XCTAssertFalse(modelText.contains("public func applySettings"), "WorkspaceModel.swift should not own settings application API bodies.")
+        XCTAssertFalse(modelText.contains("public func applyRuntime"), "WorkspaceModel.swift should not own runtime application API bodies.")
+        XCTAssertFalse(modelText.contains("public func setAgentStatus"), "WorkspaceModel.swift should not own agent status override API bodies.")
         XCTAssertFalse(modelText.contains("TrustedRouterDefaults.normalizedDefaultModelID(model)"), "WorkspaceModel should not own model ID normalization.")
         XCTAssertFalse(modelText.contains("root.config.favoriteModels.append"), "WorkspaceModel should not mutate favorite-model arrays directly.")
         XCTAssertFalse(modelText.contains("TrustedRouterDefaults.normalizedModelCatalog(models)"), "WorkspaceModel should not own catalog normalization.")
